admin-modules/requirements.txt
pytz==2025.2
pandas==2.2.3
pyjwt==2.10.1
httpx==0.28.1
uvicorn==0.34.3
pydantic==2.8.2
requests==2.32.3
gunicorn==23.0.0
fastapi==0.115.12
psycopg2-binary==2.9.10
python-multipart==0.0.20